Create custom mapper test and needed data objects
The test consists of two tests:
 The first, customMappedObjecTest, is used to find out whether the
(un-)marshaling works for objects mapped with a custom mapper. It will
try to save the object as a property, as part of a list and as a value
in a map.
The second, complexCustomMappingTest, goes a step further and tries to
(un-)marshal a list containing a map with a custom mapped object as
well as a map containing a list with a custom mapped object as value.
This test is used to guarantee that the mapping works over multiple
layers. The ‚complexest‘ list and map are used to prove the possibility
of recursion when mapping the objects. This may be a bit far fetched
and not fitting perfectly here. It could be put into the complex test
section somehow. Anyway this tests guarantee that custom mapped objects
will be unmapped no matter where they are.