Add support for CustomMappedObjects nested in Lists or Maps
All tests which I think are relevant passed, so here is what has
changed/how it works:
- There was some duplicate code in the createMap and createList
methods. This has been moved to unmarshallInternal.
- There is a method fillList to fill a list which further unpacks the
data. There was no such method for maps… I created one based on the
existing method. I reused unmarshallInternal as a lot of function was
basically the same. To allow nested elements the parameterized type of
the list/map has to be passed to the function. It will be unpacked to
the next level as necessary.
- As arrays are no parameterized types I created a small anonymus class
which basically works as a wrapper around the array type making it
possible to use the methods for arrays to.
- The custom mapper check for fields of an entity/embedded has been
moved. There also has been the addition so that both directly mapped
values and packed mappings are possible. It is more difficult to
describe it than it actually is (lines 692-696)

That’s all! The created tests pass and previous on other parts seems to
be preserved.