Make UuidRepresentation configurable or UUIDs can not be stored
Mongo Driver 4 has UuidRepresentation.UNSPECIFIED as default, what
renders the driver and morhpium unable to store UUID values. If tried
the driver will throw an exception. UuidRepresentation can now be
configured with MorphiumConfig. UuidRepresentation.STANDARD should be
preferred if not an other application that only supportes JAVA_LEGACY is
using the same database.

Morphium will creae UUID objects from JAVA_LEGACY (type 3) UUID values
from the database that are delivered as byte array from the driver.

UUID objects are supported as keys and as values in Maps.

Further changes: LocalDate, LocalTime and LocalDateTime are stored as
Dates in the mongo database