making null value handling more intuitive
up until now, null values from db were ignored unles @useifnull was
added. This is counter intuitive and due to bugs did not work every time
now null values are taken as they are: values. if a field is not in db -
use the default in java. there is a new annotation @IngoreNullFromDB to
ignore null values - avoid npe - treat them as non existant fields